refactor: rename component directories to camel case
