+++ src/pages/follow-up-analysis/Form.js
@@ -3,8 +3,8 @@ import i18n from '@dhis2/d2-i18n'
 import { Button } from '@dhis2/ui'
 import DatePicker from 'material-ui/DatePicker'
 import React from 'react'
-import AvailableDatasetsSelect from '../../components/available-datasets-select/AvailableDatasetsSelect'
-import AvailableOrganisationUnitsTree from '../../components/available-organisation-units-tree/AvailableOrganisationUnitsTree'
+import AvailableDatasetsSelect from '../../components/AvailableDatasetsSelect/AvailableDatasetsSelect'
+import AvailableOrganisationUnitsTree from '../../components/AvailableOrganisationUnitsTree/AvailableOrganisationUnitsTree'
 import cssPageStyles from '../Page.module.css'
 import jsPageStyles from '../PageStyles'
