+++ src/pages/follow-up-analysis/follow-up-analysis-table/FollowUpAnalysisTable.js
@@ -16,8 +16,8 @@ import {
 } from 'material-ui'
 import PropTypes from 'prop-types'
 import React, { Component } from 'react'
-import DownloadAs from '../../../components/download-as/DownloadAs'
-import FormattedNumber from '../../../components/formatters/FormattedNumber'
+import DownloadAs from '../../../components/DownloadAs/DownloadAs'
+import FormattedNumber from '../../../components/FormattedNumber/FormattedNumber'
 import { apiConf } from '../../../server.conf'
 import cssPageStyles from '../../Page.module.css'
 import jsPageStyles from '../../PageStyles'
