+++ src/pages/outlier-detection/outlier-analysis-table/OutlierAnalysisTable.module.css
