+++ src/pages/validation-rules-analysis/ValidationRulesDetails/ValidationRulesDetails.js
@@ -3,7 +3,7 @@ import classNames from 'classnames'
 import { Dialog, FlatButton, FontIcon } from 'material-ui'
 import PropTypes from 'prop-types'
 import React from 'react'
-import FormattedNumber from '../../../components/formatters/FormattedNumber'
+import FormattedNumber from '../../../components/FormattedNumber/FormattedNumber'
 import { apiConf } from '../../../server.conf'
 import Page from '../../Page'
 import cssPageStyles from '../../Page.module.css'
