+++ src/pages/validation-rules-analysis/ValidationRulesDetails/ValidationRulesDetails.module.css
